// src/generated/protobuf/v1/response_enqueue.ts
// @ts-nocheck
/**
 * Generated by the protoc-gen-ts.  DO NOT EDIT!
 * compiler version: 7.35.1
 * source: response_enqueue.proto
 * git: https://github.com/thesayyn/protoc-gen-ts */
import * as pb_1 from "google-protobuf";
export namespace PBTypes {
    export class ResponseEnqueue extends pb_1.Message {
        #one_of_decls: number[][] = [];
        constructor(data?: any[] | {
            match_making_queue_added_time?: number;
        }) {
            super();
            pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
            if (!Array.isArray(data) && typeof data == "object") {
                if ("match_making_queue_added_time" in data && data.match_making_queue_added_time != undefined) {
                    this.match_making_queue_added_time = data.match_making_queue_added_time;
                }
            }
        }
        get match_making_queue_added_time() {
            return pb_1.Message.getFieldWithDefault(this, 1, 0) as number;
        }
        set match_making_queue_added_time(value: number) {
            pb_1.Message.setField(this, 1, value);
        }
        static fromObject(data: {
            match_making_queue_added_time?: number;
        }): ResponseEnqueue {
            const message = new ResponseEnqueue({});
            if (data.match_making_queue_added_time != null) {
                message.match_making_queue_added_time = data.match_making_queue_added_time;
            }
            return message;
        }
        toObject() {
            const data: {
                match_making_queue_added_time?: number;
            } = {};
            if (this.match_making_queue_added_time != null) {
                data.match_making_queue_added_time = this.match_making_queue_added_time;
            }
            return data;
        }
        serialize(): Uint8Array;
        serialize(w: pb_1.BinaryWriter): void;
        serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
            const writer = w || new pb_1.BinaryWriter();
            if (this.match_making_queue_added_time != 0)
                writer.writeInt64(1, this.match_making_queue_added_time);
            if (!w)
                return writer.getResultBuffer();
        }
        static deserialize(bytes: Uint8Array | pb_1.BinaryReader): ResponseEnqueue {
            const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new ResponseEnqueue();
            while (reader.nextField()) {
                if (reader.isEndGroup())
                    break;
                switch (reader.getFieldNumber()) {
                    case 1:
                        message.match_making_queue_added_time = reader.readInt64();
                        break;
                    default: reader.skipField();
                }
            }
            return message;
        }
        serializeBinary(): Uint8Array {
            return this.serialize();
        }
        static deserializeBinary(bytes: Uint8Array): ResponseEnqueue {
            return ResponseEnqueue.deserialize(bytes);
        }
    }
}